pyproject.toml
[build-system]
requires = ["flit_core >=3.7,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "project"
dynamic = ["version", "description"]
dependencies = [
    # PyTorch etc.
    "torch ~= 2.7",
    "lightning ~= 2.5",
    "torchmetrics",
    "torchvision",
    "einops",
    # General science & ml
    "numpy",
    "scipy",
    "scikit-learn",
    "pandas",
    # Plotting & visualization
    "matplotlib",
    # hydra & logging
    "hydra-core ~= 1.3",
    "hydra-submitit-launcher",
    "wandb ~= 0.19",
    # Utilities
    "tqdm",
    "rich",
    "ipython",
    "ipdb",
    "jupyterlab",
    "ipympl",
    "transformers>=4.51.3",
    "rouge>=1.0.1",
    "nltk==3.8.1",
    "datasets>=3.6.0",
    "line-profiler>=4.2.0",
    "more-itertools>=10.7.0",
]
requires-python = ">= 3.13"
license = {text = "MIT"}

[tool.ruff]
line-length = 90

[tool.ruff.lint.isort]
known-third-party = ["wandb"]
